+++ electron/preload.cjs
@@ -2,6 +2,7 @@ const { contextBridge, ipcRenderer } = require('electron');
 
 contextBridge.exposeInMainWorld('thingtimeDesktop', {
   getInfo: () => ipcRenderer.invoke('thingtime-desktop:get-info'),
+  loadUrl: (url) => ipcRenderer.invoke('thingtime-desktop:load-url', url),
   platform: process.platform,
   versions: {
     chrome: process.versions.chrome,